opt: allocate linked buffer on lazy-stack
